fix coding errors, remove code duplication from events
